Extracted the extension on Joined4 to be a BuildResult.
